requirements.txt
dj-static==0.0.6
Django==3.0.3
django-crispy-forms==1.9.2
django-taggit==1.2.0
djangorestframework==3.11.2
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
static3==0.7.0
whitenoise==5.2.0
wincertstore==0.2
djangorestframework==3.11.2
dj-database-url==0.5.0
pillow==8.0.1